Project5/requirements.txt
matplotlib==3.10.7
numpy==2.3.3
openpyxl==3.1.5
pandas==2.3.3
pillow==11.3.0